# internal/embedded/openapi/openapi.yaml
components:
  schemas:
...
    error:
      properties:
        error:
          $ref: '#/components/schemas/response.Error'
      type: object
...
    response.Error:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: API key for authentication (optional, configurable)
      in: header
      name: X-API-Key
      type: apiKey
...
info:
  contact:
    name: Starmap Project
    url: https://github.com/agentstation/starmap
  description: |-
    REST API for the Starmap AI model catalog with reactive catalog publication updates over SSE.

    Features:
    - Complete model and provider queries
    - Advanced filtering and search
    - Heartbeat-enabled catalog publication updates via Server-Sent Events
    - In-memory caching for performance
    - Rate limiting and authentication support
  license:
    name: MIT
    url: https://github.com/agentstation/starmap/blob/main/LICENSE
  title: Starmap API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/providers:
    get:
      description: List all providers
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/error'
                properties:
                  data:
                    type: object
                  error:
                    $ref: '#/components/schemas/response.Error'
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/error'
                properties:
                  data:
                    type: object
                  error:
                    $ref: '#/components/schemas/response.Error'
                type: object
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: List providers
      tags:
      - providers
...
  /api/v1/stats:
    get:
      description: Get complete server and catalog statistics
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/error'
                properties:
                  data:
                    type: object
                  error:
                    $ref: '#/components/schemas/response.Error'
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/error'
                properties:
                  data:
                    type: object
                  error:
                    $ref: '#/components/schemas/response.Error'
                type: object
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Catalog statistics
      tags:
      - admin
  /api/v1/update:
    post:
      description: Accept an asynchronous catalog synchronization
      parameters:
      - description: Update specific provider only
        in: query
        name: provider
        schema:
          type: string
      - description: Update one source only (local_catalog, providers, models_dev_http,
          or models_dev_git)
        in: query
        name: source
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/error'
                properties:
                  data:
                    type: object
                  error:
                    $ref: '#/components/schemas/response.Error'
                type: object
          description: Accepted
        "500":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/error'
                properties:
                  data:
                    type: object
                  error:
                    $ref: '#/components/schemas/response.Error'
                type: object
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Trigger catalog update
      tags:
      - admin
...
  /api/v1/updates/stream:
    get:
      description: Heartbeat-enabled stream of post-commit catalog publication hints
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: Event stream
      security:
      - ApiKeyAuth: []
      summary: SSE updates stream
      tags:
      - updates
servers:
- description: Local development server
  url: http://localhost:8080/api/v1